+++ projects/test/index.js
@@ -1,104 +1,8 @@
-/* ДЗ 1 - Функции */
+import './index.html';
 
-/*
- Задание 1:
- 1.1: Добавьте к функции параметр с любым именем
- 1.2: Функция должна возвращать аргумент, переданный ей в качестве параметра
- Пример:
-   returnFirstArgument(10) вернет 10
-   returnFirstArgument('привет') вернет `привет`
- Другими словами: функция должна возвращать в неизменном виде то, что поступает ей на вход
- */
-const returnFirstArgument = (arg) => arg;
+const button = document.querySelector('#my-button');
+const result = document.querySelector('#my-result');
 
-/*
- Задание 2:
- 2.1: Функция должна возвращать сумму переданных аргументов
- Пример:
-   sumWithDefaults(10, 20) вернет 30
-   sumWithDefaults(2, 4) вернет 6
- 2.2 *: Значение по умолчанию для второго аргумента должно быть равно 100
- Пример:
-   sumWithDefaults(10) вернет 110
- */
-const sumWithDefaults = (arg1, arg2 = 100) => arg1 + arg2;
-
-/*
- Задание 3:
- Функция должна принимать другую функцию и возвращать результат вызова этой функции
- Пример:
-   returnFnResult(() => 'привет') вернет 'привет'
- */
-const returnFnResult = (fn) => fn();
-
-/*
- Задание 4:
- Функция должна принимать число и возвращать новую функцию (F)
- При вызове функции F, переданное ранее число должно быть увеличено на единицу и возвращено из F
- Пример:
-   var f = returnCounter(10);
-   console.log(f()); // выведет 11
-   console.log(f()); // выведет 12
-   console.log(f()); // выведет 13
- */
-function returnCounter(number = 0) {
-  return function f() {
-    return ++number;
-  };
-}
-/*
- Задание 5 *:
- Функция должна возвращать все переданные ей аргументы в виде массива
- Количество переданных аргументов заранее неизвестно
- Пример:
-   returnArgumentsArray(1, 2, 3) вернет [1, 2, 3]
- */
-function returnArgumentsArray() {
-  const result = [];
-  for (let i = 0; i < arguments.length; i++) {
-    result[i] = arguments[i];
-  }
-  return result;
-}
-
-/*
- Задание 6 *:
- Функция должна принимать другую функцию (F) и некоторое количество дополнительных аргументов
- Функция должна привязать переданные аргументы к функции F и вернуть получившуюся функцию
- Пример:
-   function sum(a, b) {
-     return a + b;
-   }
-   var newSum = bindFunction(sum, 2, 4);
-   console.log(newSum()) выведет 6
- */
-
-// function fn(...args) {
-//   const result = [];
-//   for (let i = 0; i < args.length; i++) {
-//     result[i] = arguments[i];
-//   }
-//   return result;
-// }
-//
-// function bindFunction(fn, ...args) {
-//
-//   return function newSum() {
-//     return this.fn(...args);
-//   };
-// }
-
-function bindFunction(func, ...args) {
-  return function newSum() {
-    return func(...args);
-  };
-}
-
-export {
-  returnFirstArgument,
-  sumWithDefaults,
-  returnArgumentsArray,
-  returnFnResult,
-  returnCounter,
-  bindFunction,
-};
+button.addEventListener('click', () => {
+  result.textContent = 42;
+});
